requirements.txt
flask
azure-storage-blob
azure-cosmos